Store all components data to redux
- footerLinksSlice : Footer component data
- pricingsSlice : Pricelist component data
- serviceSlice : Service component data
- /testimonial : Testimonial component data
