+++ package.json
@@ -9,9 +9,11 @@
     "tailwind": "postcss src/styles/tailwind.css -o src/styles/index.css --watch"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
+    "react-redux": "^7.2.8",
     "react-router-dom": "^6.3.0"
   },
   "devDependencies": {
