+++ src/main.jsx
@@ -1,14 +1,18 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
+import store from './app/store'
+import { Provider } from 'react-redux'
 import { BrowserRouter } from 'react-router-dom'
 import "./styles/index.css"
 import App from './App'
 
 ReactDOM.render(
   <React.StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
+    <Provider store={store}>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </Provider>
   </React.StrictMode>,
   document.getElementById('root')
 )
